refactor(analytics): extract timezone helpers to lib/timezone with formatter cache
Addresses Copilot review feedback on PR #99:

1. `Intl.DateTimeFormat` is now cached per zone via a module-level Map.
   `aggregateDailyLoads` calls `dayInTimezone` once per activity, so
   constructing a fresh formatter on every call is wasted work — for an
   athlete with 4000+ logged activities it would reconstruct the
   formatter every single time. The formatter is stateless w.r.t. the
   input `Date`, so caching is safe.

2. The helpers move from `router/analytics.ts` to a new
   `lib/timezone.ts`. Tests previously had to import from the analytics
   router module, which transitively pulls in the DB schema and engine
   imports just to exercise three pure date helpers. The new location
   keeps the unit-test surface narrow.

No behavior change. All 9 timezone tests still pass; `@acme/api` builds
and `@acme/nextjs` typechecks clean.
