+++ packages/api/src/router/analytics.ts
@@ -24,63 +24,12 @@ import {
   predictRaceTimesFromVO2max,
 } from "@acme/engine";
 
+import { dayInTimezone, shiftIsoDay, todayInTimezone } from "../lib/timezone";
 import { protectedProcedure } from "../trpc";
 
 function getDateString(daysAgo: number): string {
   const d = new Date();
   d.setDate(d.getDate() - daysAgo);
-  return d.toISOString().split("T")[0]!;
-}
-
-/**
- * Project a `Date` onto a calendar day in a specific IANA timezone.
- *
- * Returns an ISO `YYYY-MM-DD` string. We resolve via `Intl.DateTimeFormat`
- * (zero-dependency, ships with V8) so a workout finished at 23:30 UTC on
- * day N appears on day N+1 for an athlete in Sydney and on day N for an
- * athlete in San Francisco. Falls back to UTC if the zone is invalid.
- *
- * Exported for testability — internal callers go via `aggregateDailyLoads`.
- */
-export function dayInTimezone(
-  date: Date,
-  tz: string | null | undefined,
-): string {
-  const zone = tz && tz.length > 0 ? tz : "UTC";
-  try {
-    const fmt = new Intl.DateTimeFormat("en-CA", {
-      timeZone: zone,
-      year: "numeric",
-      month: "2-digit",
-      day: "2-digit",
-    });
-    // en-CA renders as YYYY-MM-DD natively
-    return fmt.format(date);
-  } catch {
-    // Invalid TZ string — degrade to UTC rather than crashing the request.
-    return date.toISOString().split("T")[0]!;
-  }
-}
-
-/**
- * Compute "today" in the athlete's timezone as a `YYYY-MM-DD` string.
- * Used as the upper bound when zero-padding the daily-load series.
- */
-function todayInTimezone(tz: string | null | undefined): string {
-  return dayInTimezone(new Date(), tz);
-}
-
-/**
- * Step a `YYYY-MM-DD` string back by N days. Operates on the date string
- * directly so it is timezone-stable (no DST spring-forward bug).
- *
- * Exported for testability.
- */
-export function shiftIsoDay(isoDay: string, deltaDays: number): string {
-  // Anchor at noon UTC to dodge DST midnight ambiguities, then shift in
-  // milliseconds. Output is the bare YYYY-MM-DD prefix only.
-  const d = new Date(`${isoDay}T12:00:00Z`);
-  d.setUTCDate(d.getUTCDate() + deltaDays);
   return d.toISOString().split("T")[0]!;
 }
